# reborn_rentals_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Proyecto X",
        "description": "Documentación de la API con autenticación JWT (Bearer).",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Listar contactos (solo admin)",
                "description": "GET /api/contacts\nProtegido: solo ADMIN puede listar.",
                "operationId": "f048e8911c1179d582ccfc371a498d78",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Búsqueda por nombre/apellido/teléfono",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de contactos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Contact"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 32
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contact": {
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Crear contacto (público)",
                "description": "POST /api/contact\nPúblico: cualquier usuario puede crear un contacto.",
                "operationId": "ef87352f3aced2578ea632fa6ea5901d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Pérez"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+59170000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/contact/{id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Ver contacto por ID (solo admin)",
                "description": "GET /api/contact/{id}\nProtegido: solo ADMIN puede ver un contacto.",
                "operationId": "878cd8f1edc589bc6a1ec34a904f6b55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 7
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Actualizar contacto (solo admin)",
                "description": "PUT /api/contact/{id}\nProtegido: solo ADMIN puede actualizar.",
                "operationId": "25ae0312e4499122b537a1e0ac716a75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 7
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Pérez"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+59170000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Eliminar contacto (solo admin)",
                "description": "DELETE /api/contact/{id}\nProtegido: solo ADMIN puede eliminar.",
                "operationId": "3f6ad823c8115e55d154544f0299d934",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 7
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contact": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+59170000000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envia el token como: Bearer {tu_token_jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}